Add Loader and break simplifiy component Filters, creating new component FilterCardItem and pass single filter as props. Fix styles
